Added ability to read in static physics meshes from GL VBO/IBO
Functionility is contained in RigidbodyFactory as createBvhTriangleMeshFromFiles. StaticCube now uses this functionality to read in from the Assets/StaticCube directory and create its cube hitbox that way.
Also added Asset::copyBuffer function for instances when an Asset's contents need to be read after the Asset instance is destroyed. Renamed Asset::getSize to Asset::getBytes and updated uses accordingly.