Began to implement level loading functionality.
Level directories can now be parsed recursively for GameObject construction info. This info will be used to actually create the scene GameObject and all of its children.

Changed GameObject to use a string for identifiers rather than an int.

Added a new constructor to Asset which takes a std::filesystem::path and reads in the Asset from that path, rather than under the Assets folder.