Basic text loading and rendering functionality in place
Added FreeType2 as a submodule and added a TextManager class which uses FreeType2 to load a ttf font and convert it to GL textures, enabling the creation of textures from strings for use in MenuElements. An example is currently in use on the Main Menu.

Also pressing escape from the pause menu now unpauses the game cause that was annoying me that we hadn't done that yet.