Very basic level implementation.
Added an abstract Level class which acts as something of a factory for levels in the game. Implementations of Level are passed to the InGameState and contain the scene with added hardcoded elements, goal positions, etc. Future commits will add more configuration to the levels.

Created TowersLevel and CastleLevel, and removed some old level Asset files.